Make sure we interpret bool and bool announced as string the same way
Some yml files for creating the flavor are announcing some field values
like 'true' and not true. this change makes sure we are correctly
interpreting them as bool's and not as strings. So practically both are now valid,
either a bool or a string like 'true' will now work fine.

fixes #42
Signed-off-by: Robin van der Linden <linden@b1-systems.de>